Update code based on scan checks (#47)
* fix generate ir script for bsdle issue

* fix dir creation ftest logs

* assert issues fixed for sdle scan

* pip issue fixed for sdle scan

* Updated assert condition for Path

+++ test/python/scripts/tf_unittest_runner.py
@@ -21,7 +21,11 @@ import tf_test_modify_input as modify
 try:
     import xmlrunner
 except:
-    os.system('pip install unittest-xml-reporting')
+    try:
+        from pip import main as pipmain
+    except:
+        from pip._internal.main import main as pipmain
+    pipmain(['install', "unittest-xml-reporting"])
     import xmlrunner
 
 os.environ['OPENVINO_TF_DISABLE_DEASSIGN_CLUSTERS'] = '1'
@@ -231,7 +235,7 @@ def list_tests(module_list, regex_input):
 
 
 def read_tests_from_file(filename):
-    assert os.path.exists(filename),"Couldn't find the path {}".format(filename)
+    if not os.path.exists(filename): raise AssertionError("Could not find the path {}".format(filename))
     with open(filename) as list_of_tests:
         return [
             line.split('#')[0].rstrip('\n').strip(' ')
@@ -260,7 +264,7 @@ def run_test(test_list, xml_report, verbosity=0):
         for test in test_list:
             names = loader.loadTestsFromName(test)
             suite.addTest(names)
-        assert os.path.exists(xml_report), "Could not find the path"
+        if not os.path.exists(xml_report): raise AssertionError("Could not find the path")
         with open(xml_report, 'wb') as output:
             sys.stdout = open(os.devnull, "w")
             sys.stderr = open(os.devnull, "w")
